Refactored / moved around some files
