+++ backend/job_clustering/embed_jobs.py
@@ -4,16 +4,11 @@ from pathlib import Path
 import sys
 import asyncio
 
-def setup_backend_imports():
+def setup_backend_imports(path="backend"):
 	# Ensure backend/ is on sys.path so its modules import as top-level modules
 	root = Path(__file__).resolve().parents[2]
-	backend_dir = root / "backend"
+	backend_dir = root / path
 	sys.path.insert(0, str(backend_dir))
-
-# def embed_job_descriptions(job_descriptions: list[str], EMBEDDING_MODEL) -> np.ndarray:
-#     model = SentenceTransformer(EMBEDDING_MODEL)
-#     embeddings = model.encode(job_descriptions, show_progress_bar=True, normalize_embeddings=True)
-#     return np.array(embeddings)
 
 async def save_job_embeddings(job_ids: list[int], embeddings: np.ndarray, model_version: str, db_session):
     # Use endpoint in main.py to save embeddings
@@ -37,8 +32,8 @@ async def main():
         import database
         import models
         from config import EMBEDDING_MODEL
-        sys.path.append(str(Path(__file__).resolve().parents[2] / "backend"))
-        from services.sbert_embedder import embed_job_descriptions
+        setup_backend_imports("backend/services")
+        from sbert_embedder import SBERTEmbeddingService
     except Exception as e:
         print("Exception importing backend modules:", e)
 
@@ -54,7 +49,8 @@ async def main():
         job_descriptions = [jp.desc_sbert for jp in job_postings]
 
         print(f"Embedding {len(job_descriptions)} job descriptions...")
-        embeddings = embed_job_descriptions(job_descriptions)
+        embedding_service = SBERTEmbeddingService()
+        embeddings = embedding_service.embed(job_descriptions)
 
         print("Saving embeddings to database...")
         # await save_job_embeddings(job_ids, embeddings, EMBEDDING_MODEL, db_session)
